printing history

+++ app/src/main/java/LoanVariables.java
@@ -2,7 +2,7 @@ import java.util.Date;
 
 public record LoanVariables(Date startDate,
                             Date endDate,
-                            Long loan,
+                            Double loan,
                             String currency,
                             Double baseInterestRate,
                             Double marginRate) {
